error regarding objects
I was having trouble with the two tables.  The error was "cannot find column_six in table [new table]"  or something close.  I struggled with this and then decided it was the objects.  I added another object and it all started working.  
I am trying to go to sleep and realize that this makes no sense;  needing two connection objects and two cursor objects.  I went back the next morning and tried it again and .yes. the same object works for both tables.  I have no idea why it started working, but "whatever"
I propose that this is a nice little microcosm about programming.