next day is working
There has been progress.  I am able to make the SELECT according to the date.:
SELECT *  WHERE  datefield.  # this uses a subsitiution .  But it is all strings.  The python takes dates the datetime value and, in putting it                         in the sqlite3 database,   turns into a string.  But a quantitive comparison with another string works.!  Okay.

To increment to the next day.          Make the string into a datetime
                                                         Increment the day
                                                         Turn the datetime back into a string
                                                          Chop off the last 16 digits leaving only the year, month and  day.
And then the WHERE will move to the next day.
It is pretty cool.  It is great to see this working.