I am trying out the writing and formating syntax.
I have made steady progress.  This is very different than the old world of fortran, Basic and even Perl.  
There has been progress.  I am able to make the SELECT according to the date.:
SELECT *  WHERE  datefield.  # this uses a subsitiution .  But it is all strings.  The python takes dates the datetime value and, in putting it in the sqlite3 database,   turns into a string.  But a quantitive comparison with another string works.!  Okay.
To increment to the next day.  
*Make the string into a datetime
*Increment the day
*Turn the datetime back into a string
*chop off the extra characters  my_string = my_string[0:10]

Late breaking news.  The While statement responds as a Boolean.  
•	While (var_one):   var_one is the unix time as a string.  This is initialized to None and if it is not returned to a string with the fetchmany, the while statement returns zero.  So this stops the cycle at the end of the database.