This is the first prototype and works okay.
The databases are added one at a time, (a pain) and the start date of the database must match the start date of the database. (also a Pain)  This python script is using the newer databases that are not sorted in the database.  To run this in GitHub, the database name would be changed to match something in the same folder.  A person wondering about the Python/sqlite3 platform might want to look at this, .  Probably  beginner/intermediate in complexity.