Jan 22 2020.  The first Beta prototype is complete
  It works as designed.  Things I have learned:
Much of the code relates to exceptions;  what if the field is null.  What if the days are not sequential.  How about a zero value.  How can I get rid of the garbage right at the beginning.
The sqlite3 table includes a [date] field.  I just learned this.  The database was taking the datetime.datetime, but making it into a string to fit in my TEXT box.  To increment the day I had to turn it into a datetime, increment it, surn it back into a string and chop it down to the correct 10 digits.
The two pairs of objects and two databases were the usual set of carelessness and bad thinking,, but really no problem.
I am much dissappointed with the dta I have seen.  It is too sloppy to make any conclusions.