Add files via upload
April 27, 2020  I have been working on the "gravity project" and making progress.  Regarding the experiment, I have added the question:  
Is mass constant?  The handle is a rotating pendulum.  I like this approach because it plugs right into the existing plan.  So changes in rotational momentum translate to changes in mass.  I am better and better with Python and probably not a beginner anymore.