Add files via upload
This project is created by a self taught python hobbyist.  This tells you that it may be full of flaws.  
 It also says that it may not be as difficult as many repositories.  There is not much in the way of libraries 
(the ssqlite3 library, of course, and a few others).   Are you interested in; 

 Passing data from one database to another.  
Moving around within a database using the datetime (UNIX).  
Data types,
 logic statements, 
for and if and while. 
 Counters and formatting print statements.  
The main program is long but you may find it more understandable than many.  To dive in just download the files 
and run the Python program.    It looks for the database, but you just downloaded it.  
It creates a summary database with a table and puts the summary of 350K rows into 40 rows.  Pretty cool.
What have I learned:  
It is the same with science  ...   the trick is to understand the question!  
As I gained a better understanding of my objectives, the system analysis became more clear.